refactor cache spawner
