+++ executorlib/cache/executor.py
@@ -2,9 +2,9 @@ import os
 from typing import Optional
 
 from executorlib.base.executor import ExecutorBase
-from executorlib.cache.shared import (
+from executorlib.cache.shared import execute_tasks_h5
+from executorlib.standalone.cache.spawner import (
     execute_in_subprocess,
-    execute_tasks_h5,
     terminate_subprocess,
 )
 from executorlib.standalone.thread import RaisingThread
