+++ tests/test_cache_executor_mpi.py
@@ -6,7 +6,6 @@ import unittest
 
 try:
     from executorlib import FileExecutor
-    from executorlib.cache.shared import execute_tasks_h5, execute_in_subprocess
 
     skip_h5io_test = False
 except ImportError:
